pyproject.toml
[project]
name = "lucky"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.0",
]
